% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigendec}
\alias{eigendec}
\title{Spectral decomposition of a symetric matrix}
\usage{
eigendec(M)
}
\arguments{
\item{M}{a matrix.}
}
\description{
Eigen decomposition of dense symmetric/hermitian matrix M
using divide-and-conquer methods that provides slightly different results
than the standard method, but is considerably faster for large matrices.
}
