% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapser.R
\docType{package}
\name{gasper-package}
\alias{gasper}
\alias{gasper-package}
\title{gasper: GrAph Signal ProcEssing in R}
\description{
Graph Signal Processing in R. It provides 
    the standard operations for signal processing on graphs: 
    Graph Fourier Transform, Spectral Graph Wavelet Transform, 
    visualization tools. It also implements a data driven method
    for graph signal denoising.
}
\keyword{internal}
