% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft.R
\name{gft}
\alias{gft}
\title{Compute Graph Fourier Transform.}
\usage{
gft(L, f, U = NULL)
}
\arguments{
\item{L}{Laplacian matrix of the graph.}

\item{f}{Graph signal to analyze.}

\item{U}{Eigenvectors of the Laplacian matrix.  If NULL (default), the function
will compute the eigendecomposition of the Laplacian.}
}
\value{
\code{hatf} Graph Fourier Transform of f.
}
\description{
Computes the Graph Fourier Transform for a given graph signal f. It utilizes the eigenvectors U of the Laplacian matrix, and if U is not provided, it computes the eigendecomposition of the Laplacian.
}
\seealso{
\code{\link{igft}}
}
