% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid1.R
\docType{data}
\name{grid1}
\alias{grid1}
\title{grid1 graph.}
\format{
list of 4 elements
\describe{
\item{xy}{coordinates}
\item{sA}{adjacency matrix}
\item{dim}{numbers of rows, columns and numerically nonzero elements}
\item{info}{information about the matrix}
}
}
\usage{
grid1
}
\description{
A dataset containing grid1 graph.
}
\keyword{datasets}
