% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igft.R
\name{igft}
\alias{igft}
\title{Compute Inverse Graph Fourier Transform.}
\usage{
igft(L, hatf, U = NULL)
}
\arguments{
\item{L}{Laplacian matrix of the graph.}

\item{hatf}{Graph Fourier Transform of the signal to be inverted.}

\item{U}{Eigenvectors of the Laplacian matrix. If NULL (default), the function will compute the eigendecomposition of the Laplacian.}
}
\value{
\code{f} Original graph signal obtained from the inverse transform of hatf.
}
\description{
Computes the Inverse Graph Fourier Transform for a given transformed graph signal 'hatf'. ' It utilizes the eigenvectors U of the Laplacian matrix, and if U is not provided, it computes the eigendecomposition of the Laplacian.
}
\seealso{
\code{\link{gft}}
}
