% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_mat.R
\name{laplacian_mat}
\alias{laplacian_mat}
\title{Graph Laplacian matrix.}
\usage{
laplacian_mat(W)
}
\arguments{
\item{W}{Adjacency matrix.}
}
\value{
\code{L} (unnormalized) graph Laplacian matrix.
}
\description{
Compute the (unnormalized) graph Laplacian matrix from the adjacency matrix.
}
