% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filter.R
\name{plot_filter}
\alias{plot_filter}
\title{Plot tight-frame filters.}
\usage{
plot_filter(lmax, b, N = 1000)
}
\arguments{
\item{lmax}{Largest eigenvalues of the Laplacian matrix.}

\item{b}{Parameter that control the number of scales.}

\item{N}{Number of discretization points (by default N=1000).}
}
\description{
Plot tight-frame kernels/filters.
}
