% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Graph plot}
\usage{
plot_graph(z, size = 0.75)
}
\arguments{
\item{z}{A list containing graph data. This list must have the following components:
\itemize{
  \item{sA} - An adjacency matrix or a sparse Matrix representation of the graph.
  \item{xy} - A matrix or dataframe containing the x and y coordinates of each node in the graph.
}}

\item{size}{Numeric. Dot size for nodes. Default is 0.75.}
}
\description{
Visualizes a graph using ggplot2. It plots nodes as points and edges as segments connecting these points. If the input graph data is represented using a sparse matrix, it will be converted to a regular summary format for visualization.
If node coordinates are not provided, they are computed using spectral coordinates.
}
\details{
If node coordinates \code{xy} are not provided, they will be calculated using spectral methods. For larger graphs, this can be computationally intensive and may take significant time. Use with caution for large graphs if node coordinates are not supplied.
}
\examples{
data(grid1)
plot_graph(grid1)
}
\seealso{
\code{\link{plot_signal}}, \code{\link{spectral_coords}}
}
