% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randsignal.R
\name{randsignal}
\alias{randsignal}
\title{Generate Random Signal with Varying Regularity.}
\usage{
randsignal(eta, k, A, r)
}
\arguments{
\item{eta}{Smoothness parameter (numeric, between 0 and 1).}

\item{k}{Smoothness parameter (integer).}

\item{A}{Adjacency matrix. Must be symmetric (matrix).}

\item{r}{Optional (numeric). Corresponding to the largest eigenvalue of \code{A} in magnitude (numeric).}
}
\value{
\code{f} a numeric vector representing the output signal.
}
\description{
\code{randsignal} constructs a random signal with specific regularity properties, utilizing the adjacency matrix \eqn{A}{A} of the graph, a smoothness parameter \eqn{\eta}{eta}, and an exponent \eqn{k}{k}. The generation is carried out in sparse matrices in order to scale up.
}
\details{
The generated signal is formulated as
\eqn{f = A^k x_{\eta} / r^k}{f = A^k x_eta / r^k}
where \eqn{x_{\eta}}{x_eta} represents Bernoulli random variables, and \eqn{r}{r} is the largest eigenvalue of the matrix \eqn{A}{A}. The power \eqn{k}{k} essentially captures the influence of a node's \eqn{k}{k}-hop neighborhood in the generated signal, implying that a higher \eqn{k}{k} would aggregate more neighborhood information resulting in a smoother signal. The normalization by the largest eigenvalue ensures that the signal remains bounded. This signal generation can be related to the Laplacian quadratic form that quantifies the smoothness of signals on graphs. B y controlling the parameters \eqn{\eta}{eta} and \eqn{k}{k}, we can modulate the smoothness or regularity of the generated signal.
}
\examples{
\dontrun{
# Generate a signal with smoothness parameters eta = 0.7 and k = 3
f <- randsignal(eta = 0.7, k = 3, A = grid1$sA)
}
}
