% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{rlogo}
\alias{rlogo}
\title{R logo graph.}
\format{
A list with 2 elements:
\itemize{
  \item \code{xy}: A matrix representing the coordinates for each node in the graph.
  \item \code{sA}: Adjacency matrix.
}
}
\usage{
rlogo
}
\description{
A dataset containing a graph based on the R logo.
}
\keyword{datasets}
