% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_coords.R
\name{spectral_coords}
\alias{spectral_coords}
\title{Spectral Coordinates for Graph Drawing.}
\usage{
spectral_coords(adj_mat)
}
\arguments{
\item{adj_mat}{A symmetric adjacency matrix or sparse matrix representing an undirected graph.}
}
\value{
A matrix where each row represents the spectral coordinates of a node in the graph.
}
\description{
Calculates the spectral coordinates of a graph using the two smallest non-zero eigenvalues of the graph Laplacian.
}
\seealso{
\code{\link{plot_graph}}
}
