% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tight_frame.R
\name{tight_frame}
\alias{tight_frame}
\title{Tight frame computation.}
\usage{
tight_frame(evalues, evectors, b = 2)
}
\arguments{
\item{evalues}{Eigenvalues of the Laplacian matrix.}

\item{evectors}{Eigenvectors of the Laplacian matrix.}

\item{b}{Parameter that control the number of scales.}
}
\description{
Construct tight (Parseval) frame.
}
