% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analysis operator.}
\usage{
analysis(y, tf)
}
\arguments{
\item{y}{Graph signal to analyze.}

\item{tf}{frame coefficients.}
}
\value{
\code{coef} Transform coefficients.
}
\description{
Compute the analysis operator for coefficient y.
}
