% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betathresh.R
\name{betathresh}
\alias{betathresh}
\title{Apply Beta Threshold.}
\usage{
betathresh(y, t, beta)
}
\arguments{
\item{y}{Noisy Data.}

\item{t}{Threshold.}

\item{beta}{Thresholding type (beta=1: soft, beta=2: JS).}
}
\value{
\code{x} Filtered result.
}
\description{
Apply Beta Threshold.
}
