% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betathresh.R
\name{betathresh}
\alias{betathresh}
\title{Apply Beta Threshold.}
\usage{
betathresh(y, t, b)
}
\arguments{
\item{y}{Noisy Data.}

\item{t}{Threshold.}

\item{b}{Thresholding type (b=1: soft, b=2: JS).}
}
\value{
\code{x} Filtered result.
}
\description{
Apply Beta Threshold.
}
