% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_graph.R
\name{download_graph}
\alias{download_graph}
\title{Download sparse matrix form the SuiteSparse Matrix Collection.}
\usage{
download_graph(graphname, groupname)
}
\arguments{
\item{graphname}{Name of the graph to download.}

\item{groupname}{Name of the group that provides the graph.}
}
\value{
\code{graphname} a list of dataframe contening W and xy coordinates.
}
\description{
If coordinates are associated with the graphs,
they are automatically downloaded and added to the output. See \url{https://sparse.tamu.edu/} for the list of groups and graph names.
}
\examples{
graphname <- "grid1"
groupname <- "AG-Monien"
download_graph(graphname,groupname)
plot_graph(grid1)
}
