% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigendec}
\alias{eigendec}
\title{Spectral decomposition of a symetric matrix}
\usage{
eigendec(M)
}
\arguments{
\item{M}{a matrix.
//@param method "dc" indicates divide-and-conquer method (par).
//              "std" indicates standard method.}
}
\description{
Eigen decomposition of dense symmetric/hermitian matrix M
using standard or divide-and-conquer methods.
the divide-and-conquer method provides slightly different results
than the standard method, but is considerably faster for large matrices
}
