% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigensort.R
\name{eigensort}
\alias{eigensort}
\title{Spectral decomposition of a symetric matrix.}
\usage{
eigensort(x)
}
\arguments{
\item{x}{Symetric matrix whose spectral decomposition is to be computed.}
}
\description{
Computes eigenvalues and eigenvectors of matrices
(output sorted in increasing order).
}
\examples{
A <- matrix(1, ncol=2, nrow=2)
dec <- eigensort(A)
}
