% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full.R
\name{full}
\alias{full}
\title{Conversion of Symmetric Sparse Matrix to Full Matrix}
\usage{
full(sA)
}
\arguments{
\item{sA}{Symmetric sparse matrix, either in a sparse matrix format or in a three-column format, that needs to be converted into a full matrix.}
}
\value{
\code{A} Full matrix constructed from the symmetric sparse matrix \code{sA}.
}
\description{
\code{full} converts a symmetric sparse matrix, represented as \code{sA}, into a full matrix \code{A}.
}
\examples{
sA <- pittsburgh$sA
A <- full(sA)
}
\seealso{
\code{\link{fullup}}
}
