% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullup.R
\name{fullup}
\alias{fullup}
\title{Convert symmetric sparse matrix to full matrix.}
\usage{
fullup(sA)
}
\arguments{
\item{sA}{Sparse upper triangular matrix matrix to convert.}
}
\description{
Convert a symmetric sparse matrix sA stored as upper triangular matrix to full matrix A.
}
\examples{
graphname <- "grid1"
groupname <- "AG-Monien"
download_graph(graphname,groupname)
A <- fullup(grid1$sA)
}
