% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapser.R
\docType{package}
\name{gasper-package}
\alias{gasper}
\alias{gasper-package}
\title{gasper: Graph Signal Processing}
\description{
Provides the standard operations for signal processing on graphs: graph Fourier transform, spectral graph wavelet transform, visualization tools. It also implements a data driven method for graph signal denoising/regression, for details see De Loynes, Navarro, Olivier (2019) \href{https://arxiv.org/abs/1906.01882}{arXiv:1906.01882}. The package also provides an interface to the SuiteSparse Matrix Collection, \url{https://sparse.tamu.edu/}, a large and widely used set of sparse matrix benchmarks collected from a wide range of applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fabnavarro/gasper}
  \item Report bugs at \url{https://github.com/fabnavarro/gasper/issues}
}

}
\author{
\strong{Maintainer}: Fabien Navarro \email{fabien.navarro@math.cnrs.fr}

Authors:
\itemize{
  \item Basile de Loynes \email{basile.deloynes@ensai.fr}
  \item Baptiste Olivier \email{baptiste.olivier@orange.com}
}

}
\keyword{internal}
