% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapser.R
\docType{package}
\name{gasper-package}
\alias{gasper}
\alias{gasper-package}
\title{gasper: Graph Signal Processing}
\description{
Provides the standard operations for signal processing on graphs: 
    graph fourier transform, spectral graph wavelet transform, 
    visualization tools. It also implements a data driven method
    for graph signal denoising/regression.
}
\author{
\strong{Maintainer}: Fabien Navarro \email{fabien.navarro@math.cnrs.fr}

Authors:
\itemize{
  \item Basile De Loynes \email{basile.deloynes@ensai.fr}
  \item Baptiste Olivier \email{baptiste.olivier@orange.com}
}

}
\keyword{internal}
