% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid1.R
\docType{data}
\name{grid1}
\alias{grid1}
\title{Grid1 Graph from AG-Monien Graph Collection}
\format{
list of 3 elements
\itemize{
  \item \code{xy}: A matrix with the coordinates for each node in the graph.
  \item \code{sA}: A sparse matrix representation of the graph's adjacency matrix.
  \item \code{dim}: A numeric vector containing the numbers of rows, columns, and numerically nonzero elements in the adjacency matrix.
}
}
\source{
AG-Monien Graph Collection by Ralf Diekmann and Robert Preis.
}
\usage{
grid1
}
\description{
This dataset represents the "grid1" graph sourced from the AG-Monien Graph Collection,
a collection of test graphs provided by Ralf Diekmann and Robert Preis.
The AG-Monien collection encompasses graphs from various origins, including the Harwell-Boeing collection, NASA matrices, and other graphs.
}
\keyword{datasets}
