% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_mat.R
\name{laplacian_mat}
\alias{laplacian_mat}
\title{Laplacian matrix.}
\usage{
laplacian_mat(W)
}
\arguments{
\item{W}{Adjacency matrix.}
}
\description{
Compute the (unormalized) laplacian matrix from the adjacency matrix.
}
