% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_SUREthresh.R
\name{LD_SUREthresh}
\alias{LD_SUREthresh}
\title{Level Dependent Stein's Unbiased Risk Estimate.}
\usage{
LD_SUREthresh(
  J,
  wcn,
  diagWWt,
  beta = 2,
  sigma,
  hatsigma = NA,
  policy = "uniform",
  keepSURE = FALSE
)
}
\arguments{
\item{J}{Finest-scale (highest frequency).}

\item{wcn}{Noisy wavelet coefficents.}

\item{diagWWt}{Weights.}

\item{beta}{Thresholding type (beta=1: soft, beta=2: JS).}

\item{sigma}{Sd of the noise.}

\item{hatsigma}{Estimator of the sd (if any).}

\item{policy}{Dependent or uniform.}

\item{keepSURE}{Boolean allowing to export the dataframe outuput of SUREthresh (FALSE by default).}
}
\value{
\code{wcLDSURE} Wavelet coefficient estimates applying SURE scale by scale.

\code{wcLDhatSURE} if hatsigma provided, wavelet coefficient estimates applying SURE scale by scale.
}
\description{
Level Dependent adaptive threshold selection using SURE.
}
\references{
de Loynes, B., Navarro, F., Olivier, B. (2021). Data-driven thresholding in denoising with Spectral Graph Wavelet Transform. Journal of Computational and Applied Mathematics, Vol. 389.

Stein, C. M. (1981). Estimation of the mean of a multivariate normal distribution. The annals of Statistics, 1135-1151.
}
\seealso{
\code{\link{SUREthresh}}
}
