% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NYCdata.R
\docType{data}
\name{NYCdata}
\alias{NYCdata}
\title{NYC network.}
\format{list of 2 elements
\describe{
\item{A}{NYC adjacency matrix (built using exponential weights between two nodes).}
\item{f}{median price observed from miles travelled to the given drop off point.}
}}
\source{
\url{https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2018-01.csv}.
}
\usage{
NYCdata
}
\description{
A dataset containing the NYC network.
}
\keyword{datasets}
