% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Graph plot}
\usage{
plot_graph(z, size = 0.75)
}
\arguments{
\item{z}{Graph data.}

\item{size}{Dot size.}
}
\description{
Graph plot
}
\examples{
plot_graph(minnesota)
}
