% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot a signal on top of a given graph}
\usage{
plot_signal(z, f, size = 0.75, limits = range(f))
}
\arguments{
\item{z}{Graph data.}

\item{f}{Signal to plot.}

\item{size}{Dot size.}

\item{limits}{Set colormap limits.}
}
\description{
Plot a signal on top of a given graph
}
\examples{
f <- rnorm(2642)
plot_signal(minnesota, f)
}
