% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSNR.R
\name{PSNR}
\alias{PSNR}
\title{Compute the Peak Signal to Noise Ratio}
\usage{
PSNR(x, y)
}
\arguments{
\item{x}{Numeric vector/matrix. Original reference signal/image.}

\item{y}{numeric vector/matrix. Restored or noisy signal/image.}
}
\value{
\code{PSNR} Numeric. Peak Signal to Noise Ratio.
}
\description{
\code{PSNR} function computes the Peak Signal to Noise Ratio (PSNR) between two signals or images.
}
\details{
Higher values of PSNR indicate closer similarity between the original and the compared signal or image.

The PSNR is defined by:
\deqn{\mathrm{PSNR}(x,y) = 10 \log_{10}\left(\frac{\max(\max(x),\max(y))^2}{\mathrm{MSE}(x, y)}\right)}{PSNR(x,y) = 10*log10( max(max(x),max(y))^2 / mean((x-y)^2) )}
}
\examples{
x <- cos(seq(0, 10, length=100))
y <- x + rnorm(100, sd=0.5)
PSNR(x, y)
}
\seealso{
\code{\link{SNR}}
}
