% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randsignal.R
\name{randsignal}
\alias{randsignal}
\title{Generate random signal with varying regularity.}
\usage{
randsignal(eta, k, A)
}
\arguments{
\item{eta}{Smoothness parameter.}

\item{k}{Smoothness parameter.}

\item{A}{Adjacency matrix.}
}
\value{
\code{f} output signal.
}
\description{
Generate f  = A^k * x_eta / lambda_1^k,
with A the adjacency matrix and
x_eta realization of Bernoulli random variables of parameter eta.
}
