% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlogo.R
\docType{data}
\name{rlogo}
\alias{rlogo}
\title{R logo graph.}
\format{list of 2 elements
\describe{
\item{xy}{coordinates}
\item{sA}{adjacency matrix}
}}
\usage{
rlogo
}
\description{
A dataset containing a graph based on the R logo.
}
\keyword{datasets}
