% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothmodulus.R
\name{smoothmodulus}
\alias{smoothmodulus}
\title{Modulus of smoothness.}
\usage{
smoothmodulus(f, A)
}
\arguments{
\item{f}{Signal.}

\item{A}{Adjacency matrix (sparse or dense).}
}
\description{
Compute the modulus of smoothness of a graph signal.
}
\examples{
data(grid1)
A <- grid1$A
x <- grid1$xy[ ,1]
f <- sin(x)
smoothmodulus(f, A)
}
