% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNR.R
\name{SNR}
\alias{SNR}
\title{Signal to Noise Ratio.}
\usage{
SNR(x, y)
}
\arguments{
\item{x}{Original reference signal.}

\item{y}{Restored or noisy signal.}
}
\value{
Signal to Noise ratio.
}
\description{
Signal to Noise Ratio.
}
\examples{
x <- cos(seq(0, 10, length=100))
y <- x + rnorm(100, sd=0.5)
SNR(x, y)
}
