% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE_MSEthresh.R
\name{SURE_MSEthresh}
\alias{SURE_MSEthresh}
\title{Stein's Unbiased Risk Estimate.}
\usage{
SURE_MSEthresh(
  wcn,
  wcf,
  thresh,
  diagWWt,
  beta = 2,
  sigma,
  hatsigma = NA,
  policy = "uniform",
  keepwc = TRUE
)
}
\arguments{
\item{wcn}{Noisy wavelet coefficients.}

\item{wcf}{True wavelet coefficients.}

\item{thresh}{Threshold values.}

\item{diagWWt}{Weights.}

\item{beta}{Threshold type (e.g, beta=1: soft, beta=2: JS).}

\item{sigma}{Standard deviation of the noise.}

\item{hatsigma}{Estimator of the standard deviation (if any).}

\item{policy}{Dependent or uniform.}

\item{keepwc}{Boolean allowing to export the coefficients of the frame after thresholding (TRUE by default).}
}
\value{
\code{res} a dataframe containing MSE, SURE, hatSURE and their respective minima and corresponding optimal thresholds.
}
\description{
Adaptive Threshold Selection Using Principle of SURE
(The irreducible variance term is not included,
 it does not change the position of the minimum).
}
\details{
Note:
 - the calculation of the MSE is also included for comparison purpose.
}
\references{
de Loynes, B., Navarro, F., Olivier, B. (2021). Data-driven thresholding in denoising with Spectral Graph Wavelet Transform. Journal of Computational and Applied Mathematics, Vol. 389.
}
\seealso{
\code{\link{SUREthresh}}
}
