% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUREthresh.R
\name{SUREthresh}
\alias{SUREthresh}
\title{Stein's Unbiased Risk Estimate.}
\usage{
SUREthresh(wcn, tresh, diagWWt, b, sigma, hatsigma, policy)
}
\arguments{
\item{wcn}{Noisy wavelet coefficents.}

\item{tresh}{Threshold values.}

\item{diagWWt}{Weights.}

\item{b}{Thresholding type (b=1: soft, b=2: JS).}

\item{sigma}{Sd of the noise.}

\item{hatsigma}{Estimator of the sd (if any).}

\item{policy}{Dependent or uniform.}
}
\value{
\code{res} a dataframe contening SURE, hatSURE and their respective minima.
}
\description{
Adaptive Threshold Selection Using Principle of SURE
(The irreductible variance term is not included,
 it does not change the position of the minimum).
}
\references{
de Loynes, B., Navarro, F., Olivier, B. (2019). Data-driven Thresholding in Denoising with Spectral Graph Wavelet Transform. arXiv preprint arXiv:1906.01882.
}
