% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{synthesis}
\alias{synthesis}
\title{Synthesis operator.}
\usage{
synthesis(coeff, tf)
}
\arguments{
\item{coeff}{Transform coefficients.}

\item{tf}{Frame coefficients.}
}
\value{
\code{y} synthesis signal.
}
\description{
Compute the synthesis operator for coefficient coeff.
}
