% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tight_frame.R
\name{tight_frame}
\alias{tight_frame}
\title{Tight frame computation.}
\usage{
tight_frame(evalues, evectors, b = 2)
}
\arguments{
\item{evalues}{Eigenvalues of the Laplacian matrix.}

\item{evectors}{Eigenvectors of the Laplacian matrix.}

\item{b}{Parameter that control the number of scales.}
}
\description{
Construct tight (Parseval) frame.
}
\references{
Göbel, F., Blanchard, G., von Luxburg, U. (2018). Construction of tight frames on graphs and application to denoising. In Handbook of Big Data Analytics (pp. 503-522). Springer, Cham.

de Loynes, B., Navarro, F., Olivier, B. (2019). Data-driven Thresholding in Denoising with Spectral Graph Wavelet Transform. arXiv preprint arXiv:1906.01882.
}
