% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetav.R
\name{zetav}
\alias{zetav}
\title{Evaluates tight-frame kernel functions}
\usage{
zetav(x, k, b)
}
\arguments{
\item{x}{Support to evaluate the kernel (vector).}

\item{k}{Scale index (scalar).}

\item{b}{Parameter that control the number of scales (scalar).}
}
\description{
Evaluates kernel associated with the particular (Littlewood-Paley type) tight-frame construction, based on a unity partition, proposed in the reference papers.
}
\references{
Coulhon, T., Kerkyacharian, G., & Petrushev, P. (2012). Heat kernel generated frames in the setting of Dirichlet spaces. Journal of Fourier Analysis and Applications, 18(5), 995-1066.

Göbel, F., Blanchard, G., von Luxburg, U. (2018). Construction of tight frames on graphs and application to denoising. In Handbook of Big Data Analytics (pp. 503-522). Springer, Cham.

de Loynes, B., Navarro, F., Olivier, B. (2021). Data-driven thresholding in denoising with Spectral Graph Wavelet Transform. Journal of Computational and Applied Mathematics, Vol. 389.
}
